pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raspilapse"
version = "1.1.0"
description = "A Python library for creating timelapses with Raspberry Pi Camera"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["raspberry-pi", "camera", "timelapse", "photography", "picamera2"]
authors = [
    {name = "Terje Nesthus", email = "terje@ekstremedia.no"}
]
maintainers = [
    {name = "Terje Nesthus", email = "terje@ekstremedia.no"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Capture :: Digital Camera",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "pylint>=2.15.0",
    "mypy>=1.0.0",
    "packaging>=21.0",
]

[project.urls]
"Homepage" = "https://github.com/ekstremedia/raspilapse"
"Bug Tracker" = "https://github.com/ekstremedia/raspilapse/issues"
"Documentation" = "https://github.com/ekstremedia/raspilapse#readme"
"Source Code" = "https://github.com/ekstremedia/raspilapse"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "hardware: marks tests that require actual camera hardware",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
no_strict_optional = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "R0913",  # Too many arguments
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "def main\\(",
    "raise NotImplementedError",
]
